% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BE_OSM_ADMIN_COUNTRY}
\alias{BE_OSM_ADMIN_COUNTRY}
\title{BE_OSM_ADMIN_COUNTRY: Country boundaries of Belgium extracted from OpenStreetMap}
\description{
BE_OSM_ADMIN_COUNTRY: Country boundaries of Belgium extracted from OpenStreetMap.
and enriched with labels based on TF_SOC_POP_STRUCT_2015 from the BelgiumStatistics package.
Extracted as described in \code{\link{BE_OSM_ADMIN}}

\itemize{
\item COUNTRY: Belgium
}
}
\examples{
\dontrun{
data(BE_OSM_ADMIN_COUNTRY)
str(as.data.frame(BE_OSM_ADMIN_COUNTRY))

library(sp)
plot(BE_OSM_ADMIN_COUNTRY)
}
}

