% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BE_OSM_ADMIN_REGION}
\alias{BE_OSM_ADMIN_REGION}
\title{BE_OSM_ADMIN_REGION: Region boundaries of Belgium extracted from OpenStreetMap}
\description{
BE_OSM_ADMIN_REGION: Region boundaries of Belgium extracted from OpenStreetMap.
and enriched with labels based on TF_SOC_POP_STRUCT_2015 from the BelgiumStatistics package.
Extracted as described in \code{\link{BE_OSM_ADMIN}}

\itemize{
\item CD_RGN_REFNIS, TX_RGN_DESCR_NL, TX_RGN_DESCR_FR: NIS code of the region as well as NL/FR labels (based on TF_SOC_POP_STRUCT_2015)
}
}
\examples{
\dontrun{
data(BE_OSM_ADMIN_REGION)
str(as.data.frame(BE_OSM_ADMIN_REGION))

library(sp)
plot(BE_OSM_ADMIN_REGION)
}
}

