% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{BelgiumMaps.Admin-package}
\alias{BelgiumMaps.Admin}
\alias{BelgiumMaps.Admin-package}
\title{Maps with administrative boundaries (national, regions, provinces, districts, municipalities, nissections) of Belgium extracted from OpenStreetMap}
\description{
Maps with administrative boundaries (national, regions, provinces, districts, municipalities, nissections) of Belgium extracted from OpenStreetMap
}
\examples{
\dontrun{
## Administrative areas
data(BE_OSM_ADMIN)
str(as.data.frame(BE_OSM_ADMIN))

library(sp)
plot(BE_OSM_ADMIN)
plot(subset(BE_OSM_ADMIN, admin.level \%in\% "2"))
plot(subset(BE_OSM_ADMIN, admin.level \%in\% "4"))
plot(subset(BE_OSM_ADMIN, admin.level \%in\% "6"))
plot(subset(BE_OSM_ADMIN, admin.level \%in\% "7"))
plot(subset(BE_OSM_ADMIN, admin.level \%in\% "8"))
plot(subset(BE_OSM_ADMIN, admin.level \%in\% "9"))
}
}

